Refactor pkg/health
Splits the ConsulHealthChecker and related functions into its own package
"pkg/health/checker" out of "pkg/health".  This resolves a circular import
problem, wherein the "pkg/kp" package couldn't import "pkg/health" to access
the canonical health values.
